+++ jolt-core/src/zkvm/instruction/fence.rs
@@ -19,7 +19,7 @@ impl InstructionFlags for FENCE {
 }
 
 impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<FENCE> {
-    fn to_instruction_inputs(&self) -> (u64, i64) {
+    fn to_instruction_inputs(&self) -> (u64, i128) {
         (0, 0)
     }
