+++ jolt-core/src/zkvm/instruction/virtual_muli.rs
@@ -37,20 +37,20 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualMULI> 
         LookupQuery::<WORD_SIZE>::to_lookup_operands(self).1
     }
 
-    fn to_instruction_inputs(&self) -> (u64, i64) {
+    fn to_instruction_inputs(&self) -> (u64, i128) {
         match WORD_SIZE {
             #[cfg(test)]
             8 => (
                 self.register_state.rs1 as u8 as u64,
-                self.instruction.operands.imm as u8 as i64,
+                self.instruction.operands.imm as u8 as i128,
             ),
             32 => (
                 self.register_state.rs1 as u32 as u64,
-                self.instruction.operands.imm as u32 as i64,
+                self.instruction.operands.imm as u32 as i128,
             ),
             64 => (
                 self.register_state.rs1,
-                self.instruction.operands.imm as i64,
+                self.instruction.operands.imm as i128,
             ),
             _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
         }
@@ -62,7 +62,7 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualMULI> 
             #[cfg(test)]
             8 => (x as i8).wrapping_mul(y as i8) as u8 as u64,
             32 => (x as i32).wrapping_mul(y as i32) as u32 as u64,
-            64 => (x as i64).wrapping_mul(y) as u64,
+            64 => (x as i64).wrapping_mul(y as i64) as u64,
             _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
         }
     }
