+++ tracer/src/instruction/format/format_i.rs
@@ -78,7 +78,7 @@ impl InstructionFormat for FormatI {
             rs1: self.rs1,
             rs2: 0,
             rd: self.rd,
-            imm: self.imm as i64,
+            imm: self.imm as i128,
         }
     }
 }
