+++ tracer/src/instruction/format/format_u.rs
@@ -69,7 +69,7 @@ impl InstructionFormat for FormatU {
             rs1: 0,
             rs2: 0,
             rd: self.rd,
-            imm: self.imm as i64,
+            imm: self.imm as i128,
         }
     }
 }
